+++ lib/packages/wikibase/data-model/src/Entity/Entity.php
@@ -565,4 +565,14 @@ abstract class Entity implements \Comparable, ClaimAggregate, FingerprintProvide
 	 */
 	public abstract function getType();
 
+	/**
+	 * Returns if the Entity has no content.
+	 * Having an id set does not count as having content.
+	 *
+	 * @since 0.1
+	 *
+	 * @return boolean
+	 */
+	public abstract function isEmpty();
+
 }